automated_tests/requirements.txt
beautifulsoup4==4.3.2
behave-py3
fake-factory==0.5.2
nose==1.3.4
mandrill==1.0.57
PyVirtualDisplay==0.1.5
requests==2.7.0
selenium==2.46.0
temp-mail==0.2
git+git://github.com/APodobytkorps/webium#egg=webium